Выкинул базовые репозитории Теперь *Storage - просто заглушка с массивом десериализованных записей Функционал старых *Storage унёс в репозитории @lololozhkin - когда будешь делать общение с базой, тебе надо будет переписать эти репозитории
