--- src/core/baseRepositories/baseSurveysRepository.js
@@ -1,8 +1,0 @@
-const baseSurveysRepository =
-class {
-	constructor (surveysStorage){
-		this.surveysStorage = surveysStorage;
-	}
-};
-
-module.exports = baseSurveysRepository;
