--- src/core/baseRepositories/baseUserRepository.js
@@ -1,8 +1,0 @@
-const baseUserRepository =
-class {
-	constructor(userStorage) {
-		this.userStorage = userStorage;
-	}
-};
-
-module.exports = baseUserRepository;
