+++ src/infrastructure/repositories/surveysRepository.js
@@ -1,20 +1,32 @@
-const baseSurveysRepository = require('../../core/baseRepositories/baseSurveysRepository');
 const surveysStorage = require('../surveyStorage');
 
-const SurveysRepository =
-class extends baseSurveysRepository{
-	getAllSurveysByUserID(userID){
-		return this.surveysStorage.getCreatedByUser(userID);
+class SurveysRepository {
+	constructor (surveysStorage){
+		this.surveysStorage = surveysStorage;
 	}
 
-	getSurveyByID(surveyId){
-		return this.surveysStorage.getSurveyById(surveyId);
+	getCreatedByUser(user){
+		const result = [];
+		for (const survey of this.surveysStorage)
+			if (user.created.includes(survey.id))
+				result.push(survey);
+		return result;
 	}
 
-	getAllVotedSurveysByUserID(userID){
-		return this.surveysStorage.getVotedByUser(userID);
+	getSurveyById(surveyId){
+		for (const survey of this.surveysStorage)
+			if (survey.id === surveyId)
+				return survey;
 	}
-};
+
+	getAllVotedSurveysByUser(user){
+		const result = [];
+		for (const survey of this.surveysStorage)
+			if (user.voted.includes(survey.id))
+				result.push(survey);
+		return result;
+	}
+}
 
 
 module.exports = new SurveysRepository(surveysStorage);
